move all localstorage functions to be inside of PersistenceManager class
